feat: redis 제거
