+++ docker-compose/docker-compose.yml
@@ -12,8 +12,3 @@ services:
     restart: always
     volumes:
         - ./pgdata:/var/lib/postgresql/data
-
-  redis:
-    image: 'redis:alpine'
-    ports:
-      - '6379:6379'
