+++ package.json
@@ -46,7 +46,6 @@
     "@types/cache-manager-redis-store": "^2.0.0",
     "axios": "^1.4.0",
     "cache-manager": "^4.0.0",
-    "cache-manager-redis-store": "^2.0.0",
     "cheerio": "^1.0.0-rc.12",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
