--- src/redis/redis.module.ts
@@ -1,22 +1,0 @@
-import { CacheModule, Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import * as redisStore from 'cache-manager-redis-store';
-import { RedisService } from './services/redis.service';
-
-@Module({
-  imports: [
-    CacheModule.registerAsync({
-      imports: [ConfigModule],
-      useFactory: async (configService: ConfigService) => ({
-        store: redisStore,
-        host: configService.get('REDIS_HOST'),
-        port: configService.get('REDIS_PORT'),
-        ttl: 0,
-      }),
-      inject: [ConfigService],
-    }),
-  ],
-  providers: [RedisService],
-  exports: [RedisService],
-})
-export class RedisModule {}
