+++ src/visitor/service/visitor.service.ts
@@ -1,14 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { CACHE_MANAGER, Inject, Injectable } from '@nestjs/common';
 import {
   GetTodayVisitorResponseDto,
   VisitorCountUpResponseDto,
 } from '../dtos/visitor-response.dto';
 import { Request } from 'express';
-import { RedisService } from '../../redis/services/redis.service';
+import { Cache } from 'cache-manager';
 
 @Injectable()
 export class VisitorService {
-  constructor(private readonly redisService: RedisService) {}
+  constructor(@Inject(CACHE_MANAGER) private readonly cacheManager: Cache) {}
 
   /**
    * 방문자가 늘어날 때, Cache 에 User Agent 정보를 기반으로 저장합니다.
@@ -20,7 +20,7 @@ export class VisitorService {
       const userAgent = req.get('user-agent');
       const uniqueUserInfo = `visitor-${userAgent}${ip}`;
 
-      await this.redisService.set(uniqueUserInfo, 'visited');
+      await this.cacheManager.set(uniqueUserInfo, 'visited');
 
       result.Status = 'Success';
     } catch (err) {
@@ -36,8 +36,9 @@ export class VisitorService {
   async getTodayVisitor(): Promise<GetTodayVisitorResponseDto> {
     const result: GetTodayVisitorResponseDto = { Count: 0 };
     try {
-      const allVisitors = await this.redisService.getMultipleKeysPrefix(
-        'visitor',
+      const allKeys = await ((await this.cacheManager.store) as any).keys();
+      const allVisitors = allKeys.filter((key: string) =>
+        key.includes('visitor-'),
       );
 
       result.Count = allVisitors.length;
