+++ src/visitor/visitor.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { VisitorController } from './controller/visitor.controller';
 import { VisitorService } from './service/visitor.service';
-import { RedisModule } from '../redis/redis.module';
+
+import { CacheModule as NestCacheModule } from '@nestjs/common/cache/cache.module';
 
 @Module({
-  imports: [RedisModule],
+  imports: [NestCacheModule.register()],
   providers: [VisitorService],
   controllers: [VisitorController],
   exports: [VisitorService],
